renamed properties

+++ Model/DTOs/DepreciationDto.cs
@@ -2,6 +2,6 @@ namespace kalkulator.net.Model.DTOs;
 
 public class DepreciationDto
 {
-    public decimal DepreciationRate { get; set; } // "AfA Satz"
-    public decimal BuildingValuePercentageOfPurchasePrice { get; set; } // "Anteil Gebäude an Kaufpreis"
+    public double DepreciationRate { get; set; } // "AfA Satz"
+    public double BuildingValuePercentageOfPurchasePrice { get; set; } // "Anteil Gebäude an Kaufpreis"
 }
