+++ Model/DTOs/InitialInvestmentDto.cs
@@ -3,7 +3,7 @@ namespace kalkulator.net.Model.DTOs;
 public class InitialInvestmentDto
 {
     public string? Name { get; set; }
-    public decimal Cost { get; set; }
+    public double Cost { get; set; }
     public TaxTreatment TaxTreatment { get; set; }
-    public decimal ValueIncrease { get; set; }
+    public double ValueIncrease { get; set; }
 }
