+++ Model/DTOs/LoanDto.cs
@@ -2,9 +2,9 @@ namespace kalkulator.net.Model.DTOs;
 
 public class LoanDto
 {
-    public decimal LoanAmount { get; set; } // "Darlehenssumme"
-    public decimal InterestRate { get; set; } // "Zinssatz"
-    public decimal InitialRepaymentRate { get; set; } // "Anfängliche Tilgung"
-    public decimal MonthlyPayment { get; set; } // "Kapitaldienst pro Monat"
+    public double LoanAmount { get; set; } // "Darlehenssumme"
+    public double InterestRate { get; set; } // "Zinssatz"
+    public double InitialRepaymentRate { get; set; } // "Anfängliche Tilgung"
+    public double MonthlyPayment { get; set; } // "Kapitaldienst pro Monat"
     public int YearOfFullRepayment { get; set; } // "Jahr der Volltilgung"
 }
