+++ Model/DTOs/RentDto.cs
@@ -2,8 +2,8 @@ namespace kalkulator.net.Model.DTOs;
 
 public class RentDto
 {
-    public decimal ColdRentPerSquareMeter { get; set; } // "Kaltmiete Wohnfläche pro qm"
-    public decimal TotalColdRent { get; set; } // "Kaltmiete Wohnfläche gesamt"
-    public decimal ParkingSpaces { get; set; } // "Stellplätze"
-    public decimal Other { get; set; } // "Sonstiges"
+    public double ColdRentPerSquareMeter { get; set; } // "Kaltmiete Wohnfläche pro qm"
+    public double TotalColdRent { get; set; } // "Kaltmiete Wohnfläche gesamt"
+    public double ParkingSpaces { get; set; } // "Stellplätze"
+    public double Other { get; set; } // "Sonstiges"
 }
