+++ Model/InitialInvestment.cs
@@ -6,9 +6,9 @@ public class InitialInvestment
     public int CalculationId { get; set; } // Foreign key to Calculation
     public Calculation? Calculation { get; set; } // Navigation property back to Calculation
     public string? Name { get; set; }
-    public decimal Cost { get; set; }
+    public double Cost { get; set; }
     public TaxTreatment TaxTreatment { get; set; }
-    public decimal ValueIncrease { get; set; }
+    public double ValueIncrease { get; set; }
 }
 
 public enum TaxTreatment
