+++ Model/Loan.cs
@@ -3,10 +3,10 @@ namespace kalkulator.net.Model;
 public class Loan
 {
     public int Id { get; set; } // Primary key
-    public decimal LoanAmount { get; set; } // "Darlehenssumme"
-    public decimal InterestRate { get; set; } // "Zinssatz"
-    public decimal InitialRepaymentRate { get; set; } // "Anfängliche Tilgung"
-    public decimal MonthlyPayment { get; set; } // "Kapitaldienst pro Monat"
+    public double LoanAmount { get; set; } // "Darlehenssumme"
+    public double InterestRate { get; set; } // "Zinssatz"
+    public double InitialRepaymentRate { get; set; } // "Anfängliche Tilgung"
+    public double MonthlyPayment { get; set; } // "Kapitaldienst pro Monat"
     public int YearOfFullRepayment { get; set; } // "Jahr der Volltilgung"
 
     public int CalculationId { get; set; } // Foreign key to Calculation
