+++ Model/Rent.cs
@@ -8,10 +8,10 @@ namespace kalkulator.net.Model;
 public class Rent
 {
     public int Id { get; set; } // Primary key
-    public decimal ColdRentPerSquareMeter { get; set; } // "Kaltmiete Wohnfläche pro qm"
-    public decimal TotalColdRent { get; set; } // "Kaltmiete Wohnfläche gesamt"
-    public decimal ParkingSpaces { get; set; } // "Stellplätze"
-    public decimal Other { get; set; } // "Sonstiges"
+    public double ColdRentPerSquareMeter { get; set; } // "Kaltmiete Wohnfläche pro qm"
+    public double TotalColdRent { get; set; } // "Kaltmiete Wohnfläche gesamt"
+    public double ParkingSpaces { get; set; } // "Stellplätze"
+    public double Other { get; set; } // "Sonstiges"
     
     public int CalculationId { get; set; } // Foreign key to Calculation
     // Navigation property back to Calculation
